pyproject.toml
[tool.poetry]
name = "modufur"
version = "0.1.0"
description = "Modufur Discord Bot"
authors = ["Myned <devmyned@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "~3.10"
toml = "*"
uvloop = "*"
aiohttp = "*"
urlextract = "*"
tldextract = "*"
hikari = {extras = ["speedups"], version = "*"}
hikari-lightbulb = "*"
hikari-miru = "*"
pysaucenao = {git = "https://github.com/Myned/pysaucenao.git"}

[tool.poetry.dev-dependencies]
black = "*"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
